remove "暸解" (note the 日 radical) footgun
Use 瞭 instead (note the 目 radical).

暸解 is a footgun: from the frequency in essay-tw.txt we can already see
it's barely used. We can also see that the word isn't documented in the
main MOE dictionaries[1].

So when it shows up in the candidates, it just serves to confuse the
person typing.

Also, Unihan marks the pronunciation of U+66B8 暸 as liáo (ㄌㄧㄠˊ,
second tone)[2], so follow that.

[1]: https://kemdict.com/search?q=暸&m=prefix
[2]: https://www.unicode.org/cgi-bin/GetUnihanData.pl?codepoint=暸