弧秒; sort 金鑰 higher than 緊要 or 禁藥
Honestly the problem here is that RIME doesn't have a mechanism where
tone 1 words have priority when they match exactly (since normally the
first tone doesn't get written out). Or a mechanism for a schema to
define that behavior.

That's fine, we can just tune the frequencies ourselves. This adds to
the language-specific nature of the frequency list, but I do accept that
reality.