喫 -> 吃
Again, the goal of this is to use zh_TW in essay-tw.txt in general while
still allowing the "traditional" characters to be typed. If we just used
OpenCC we would ensure we always use zh_TW but the traditional
characters would also always be transformed, making it impossible to
type them.